backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic[email]==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.19
redis[hiredis]==5.2.1
celery==5.4.0
chromadb-client==0.5.23
httpx==0.28.1
langchain-text-splitters==0.3.4
pymupdf==1.25.1
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5
tiktoken==0.8.0
markdown==3.7
sse-starlette==2.2.1
structlog==24.4.0
rank-bm25==0.2.2
pyotp==2.9.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
aiosqlite==0.20.0
